+++ src/components/ProductCard.jsx
@@ -1,4 +1,8 @@
+import { useCart } from "../context/CartContext";
+
 const ProductCard = ({product}) => {
+     const {addToCart} = useCart();
+
     return ( 
          <div  className="bg-white rounded-lg shadow p-4 flex flex-col">
               <img 
@@ -9,6 +13,10 @@ const ProductCard = ({product}) => {
                 <h2 className="text-xl font-semibold">{product.name}</h2>
                 <p className="text-gray-500 text-sm mb-2">{product.description}</p>
                 <p className="font-bold text-lg">${product.price.toFixed(2) }</p>
+
+                <button onClick={() => addToCart(product)} className="bg-blue-600 text-white mt-3 px-4 py-2 rounded transition hover:bg-blue-700">
+                    Add To Cart
+               </button>
             </div>
      );
 }
