+++ src/context/CartContext.jsx
@@ -1,0 +1,8 @@
+import { createContext, useContext} from "react";
+
+export const CartContext = createContext();
+
+
+export function useCart() {
+    return useContext(CartContext);
+};
