+++ src/main.jsx
@@ -2,13 +2,15 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
-import { ProductProvider } from './context/ProductContext.jsx'
-
+import { ProductProvider } from './context/ProductProvider.jsx'
+import { CartProvider } from './context/CartProvider.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <ProductProvider>
-      <App />
+      <CartProvider>
+        <App />
+      </CartProvider>
     </ProductProvider>
   </StrictMode>,
 )
